Adds CODEOWNERS file
_This PR was auto-generated._

This adds a CODEOWNERS file to `./github/CODEOWNERS`.
The users in the CODEOWNERS file were copied from the `shepherds` in the launch YML file.

**Why CODEOWNERS?**

The CODEOWNERS file is used to auto-assign owners (we call them shepherds) as reviewers.
We hope this will help ensure that shepherds get notified about and review changes.

NOTE: To reduce burden on the shepherds, you should still assign a member of your own team as the primary reviewer.
The shepherd should be treated as a secondary reviewer, with a special role in guiding the project.

If you have questions, please ask here or in #oncall-infra. Thanks!